test: pin the enclave pull contract alongside other wire values
